requirements.txt
asgiref==3.9.1
awsebcli==3.25
bcrypt==4.3.0
blessed==1.21.0
boto3==1.39.9
botocore==1.39.9
cement==2.10.14
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==3.0.1
Django==4.2
django-cors-headers==4.7.0
django-db-connection-pool==1.2.6
django-extensions==3.2.3
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
fabric==3.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pillow==11.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
semantic-version==2.10.0
six==1.17.0
SQLAlchemy==2.0.42
sqlparams==6.2.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.14.1
urllib3==1.26.20
wcwidth==0.2.13
whitenoise==6.9.0
wrapt==1.17.2
